+++ src/program/GameThread.cpp
@@ -18,6 +18,7 @@
  */
 
 #include "GameThread.h"
+#include "AutoDetect.h"
 #include "utils.h"
 #include "Context.h"
 #include "../shared/SharedConfig.h"
@@ -27,22 +28,6 @@
 #include <iostream>
 #include <unistd.h> // chdir()
 #include <fcntl.h> // O_RDWR, O_CREAT
-
-int GameThread::detect_arch(Context *context)
-{
-    /* Change settings based on game arch */
-    int gameArch = extractBinaryType(context->gamepath);    
-    int libtasArch = extractBinaryType(context->libtaspath);
-
-    /* Switch to libtas32.so if required */
-    if ((((gameArch&BT_TYPEMASK) == BT_ELF32) || ((gameArch&BT_TYPEMASK) == BT_PE32)) && (libtasArch == BT_ELF64)) {
-        context->libtaspath = context->libtas32path;
-        /* libtas32.so presence was already checked in ui/ErrorChecking.cpp */
-        libtasArch = extractBinaryType(context->libtaspath);
-    }
-
-    return gameArch;
-}
 
 void GameThread::set_env_variables(Context *context, int gameArch)
 {
@@ -168,16 +153,9 @@ std::list<std::string> GameThread::build_arg_list(Context *context, int gameArch
             /* wine[64] presence was already checked in ui/ErrorChecking.cpp */
             std::string cmd = "which ";
             cmd += winename;
-            FILE *output = popen(cmd.c_str(), "r");
-            if (output != NULL) {
-                std::array<char,256> buf;
-                if (fgets(buf.data(), buf.size(), output) != 0) {
-                    std::string winepath = std::string(buf.data());
-                    winepath.pop_back(); // remove trailing newline
-                    arg_list.push_back(winepath);
-                }
-                pclose(output);
-            }
+            
+            std::string winepath = queryCmd(cmd);
+            arg_list.push_back(winepath);
         }
 
         /* Push the game executable as the first command-line argument */
@@ -198,13 +176,7 @@ std::list<std::string> GameThread::build_arg_list(Context *context, int gameArch
                 break;
             }
 
-            FILE *output = popen(cmd.c_str(), "r");
-            std::array<char,256> buf;
-            fgets(buf.data(), buf.size(), output);
-            std::string dbgpath = std::string(buf.data());
-            dbgpath.pop_back(); // remove trailing newline
-            pclose(output);
-
+            std::string dbgpath = queryCmd(cmd);
             arg_list.push_back(dbgpath);
 
             /* Push debugger arguments */
@@ -284,66 +256,17 @@ std::list<std::string> GameThread::build_arg_list(Context *context, int gameArch
     return arg_list;
 }
 
-void GameThread::detect_game_libraries(Context *context)
-{
-    /* Build a command to parse the first missing library from the game executable,
-     * look at game directory and sub-directories for it */
-    std::ostringstream oss_ml;
-    oss_ml << "ldd '" << context->gamepath;
-    oss_ml << "' | awk '/ => not found/ { print $1 }' | head -1";
-    
-    FILE *output = popen(oss_ml.str().c_str(), "r");
-    std::array<char,256> buf;
-    buf[0] = '\0';
-    fgets(buf.data(), buf.size(), output);
-    std::string missing_lib = std::string(buf.data());
-    pclose(output);
-
-    if (missing_lib.empty()) return;
-    
-    missing_lib.pop_back(); // remove trailing newline
-
-    std::cout << "Try to find the location of " << missing_lib << " among game files."<< std::endl;
-
-    std::string gamedir = dirFromPath(context->gamepath);
-    std::ostringstream oss_lp;
-    oss_lp << "find '" << gamedir << "' -name " << missing_lib << " -type f -print -quit";
-
-    output = popen(oss_lp.str().c_str(), "r");
-    buf[0] = '\0';
-    fgets(buf.data(), buf.size(), output);
-    std::string found_lib = std::string(buf.data());
-    pclose(output);
-    
-    if (!found_lib.empty()) {
-        found_lib.pop_back(); // remove trailing newline
-
-        std::cout << "-> library was found at location " << found_lib << std::endl;
-        
-        std::string found_lib_dir = dirFromPath(found_lib);
-
-        char* oldlibpath = getenv("LD_LIBRARY_PATH");
-        if (oldlibpath) {
-            found_lib_dir.append(":");
-            found_lib_dir.append(oldlibpath);
-        }
-        setenv("LD_LIBRARY_PATH", found_lib_dir.c_str(), 1);
-    }
-    else {
-        std::cerr << "-> could not find the library among the game files" << std::endl;
-    }
-}
-
 void GameThread::launch(Context *context)
 {
     /* Detect the game executable arch and handle 32-bit game on 64-bit arch case */
-    int gameArch = detect_arch(context);
+    int gameArch = AutoDetect::arch(context);
 
     /* Set all environment variables */
     set_env_variables(context, gameArch);
     
+    /* Tries to detect a library folder within the game directory */
 #ifdef __unix__
-    detect_game_libraries(context);
+    AutoDetect::game_libraries(context);
 #endif
 
     /* Set where stderr of the game is redirected */