+++ src/program/utils.h
@@ -61,4 +61,7 @@ int extractBinaryType(std::string path);
 /* Get the executable from MacOS .app directory. */
 std::string extractMacOSExecutable(std::string path);
 
+/* Get the result of a shell command */
+std::string queryCmd(const std::string& cmd);
+
 #endif