+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * [ImGui] Show game inside a window (OpenGL)
 * [ImGui] Show game inside a window (Vulkan)
 * Add lua memory.baseAddress() function (#561)
+* Autodetect game engine and enforce some settings
 
 ### Changed